Allow popups coming from the Gladly ad unit.
By default, ublock blocks popups where the root hostname is a blocked domain (e.g. Google's Doubleclick domains). However, when the user clicks on an ad in the Goodblock ad unit, we want to allow popups.

Here, we mark when the Goodblock ad unit is open. When a popup occurs, if it's coming from the tab with an open Goodblock ad unit, let it go through. Otherwise, block it as usual.