Remove donation links and add apology
@chrisaljoudi, please consider accepting this pull request and ceasing
your highly unmoral behavior of soliciting donations for the work of
other individuals.

uBlock is not your project in spirit, only in name, and the fact that
you openly come up with excuses to solicit donations from end users
for a project you had essentially zero influence on is not only an
insult to each and every one of your donors, but also gives free
software developers a bad name in general by providing an
outrageously low example.

In the spirit of maintaining a community which is based on mutual
respect and sharing code, rather than mislabelling stolen code for
profit, I strongly urge you to reconsider your values regarding this
subject and free software in general.